+++ Classes/Domain/Prop/Deferrable.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace ZktSn0w\Inertia\Domain\Prop;
+
+interface Deferrable
+{
+    public function shouldDefer(): bool;
+    public function group(): string;
+}